Revert attempt to select p256_sikep434 as default key exchange
Part of the change from 58860b28cae7ab700b65e2f79ee32d9d6dcdd708 ended up working by accident due to a bug in OpenVPN, which was fixed in v2.4.9. Instead of working around it, go back to conforming to OpenVPN's intended behavior of letting OpenSSL choose suitable defaults, in the expectation the OQS fork will eventually offer only post-quantum algorithms as the default. We just have to make it very clear that users must specify an ecdh-curve directive in configuration to be sure of a post-quantum key exchange for now, much as they have to supply certificates using a post-quantum digital signature algorithm for protection there.

This commit does not revert all of the aforementioned commit, just this one part.